feat: add models, schema.sql and begin services

+++ ewm-service/src/main/java/ru/practicum/ewm/base/dto/EventShortDto.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 public class EventShortDto {
     private String annotation;
     private CategoryDto category;
-    private int confirmedRequests;
+    private long confirmedRequests;
     private LocalDateTime eventDate;
     private Integer id;
     private UserShortDto initiator;
